backend/ dodan kontroler za dohvat filma iz baze

+++ IzvorniKod/backend/src/main/java/hr/unizg/fer/backend/repository/primary/GrupaZaDigitalizacijuRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GrupaZaDigitalizacijuRepository extends JpaRepository<GrupaZaDigitalizaciju, Integer> { }
+public interface GrupaZaDigitalizacijuRepository extends JpaRepository<GrupaZaDigitalizaciju, Long> { }
