+++ IzvorniKod/backend/src/main/java/hr/unizg/fer/backend/repository/primary/KorisnikRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KorisnikRepository extends JpaRepository<Korisnik, Integer> { }
+public interface KorisnikRepository extends JpaRepository<Korisnik, Long> { }
