+++ IzvorniKod/backend/src/main/java/hr/unizg/fer/backend/repository/primary/StatistikaDigitalizacijeRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatistikaDigitalizacijeRepository extends JpaRepository<StatistikaDigitalizacije, Integer> { }
+public interface StatistikaDigitalizacijeRepository extends JpaRepository<StatistikaDigitalizacije, Long> { }
